refactor(javascript): fold the SoA column store into syntax.js
Move the C1 scaffolding (SoaAst columns + facades) from its own file
into syntax.js next to the emitter seam, matching how the CSS and HTML
SoA backends live inside their syntax modules; battery tests move into
WebpackParser.unittest.js. No behavior change.
